+++ Geometry/IManifold0.h
@@ -6,7 +6,7 @@
 namespace Geometry
 {
 
-   template<typename T, int N>
+   template<typename T, size_t N>
    class IManifold0 : public IManifold<T, N>
    {
    public:
