+++ Geometry/SamePointPredicate.h
@@ -2,7 +2,7 @@
 
 #include "IGeometryPredicate.h"
 
-template<typename T, int N>
+template<typename T, size_t N>
 struct SamePointPredicate
 {
    const IGeometryPredicate<T, N>& predicate;
@@ -10,7 +10,7 @@ struct SamePointPredicate
    bool operator()(const Point<T, N>& p) const;
 };
 
-template<typename T, int N>
+template<typename T, size_t N>
 bool SamePointPredicate<T, N>::operator()(const Point<T, N>& p) const
 {
    return predicate.samePoints(referencePoint, p);
